+++ src/render/lighting.h
@@ -100,8 +100,8 @@ enum Lightings {
  */
 typedef struct LightingBuffer {
     struct Node **lightings; /**< Array of lighting nodes. */
-    u8 length; /**< Length of the array. */
-    u8 index; /**< Current index in the array. */
+    u32 length; /**< Length of the array. */
+    u32 index; /**< Current index in the array. */
 } LightingBuffer;
